Move mappings out of config and read from separate file.
